+++ Assets/Prefabs/GameManager.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ee6520f00dda4424eb6f04281e00302d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
