+++ Assets/Prefabs/_Gameplay Prefabs/Barktender.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b8b67179f41da2447b3fa8947a05d0c3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
