+++ Assets/Prefabs/_Gameplay Prefabs/Difficulty Scaler.prefab.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 8f9934ac6cdc24816b67efa290eeded7
-folderAsset: yes
-DefaultImporter:
+guid: 4d3960e15a13456498e7f2f10328bcf5
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 100100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
