+++ Assets/Prefabs/_Gameplay Prefabs/Drink.prefab
@@ -51,10 +51,12 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: caaa21a929a7545dda96e00908fb5c2b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  Whiskey: 0
-  Vodka: 0
-  Rum: 0
-  Soda: 0
-  Coke: 0
-  Vermouth: 0
+  LaneValue: 0
+  WhiskeyValue: 0
+  VodkaValue: 0
+  RumValue: 0
+  SodaValue: 0
+  CokeValue: 0
+  VermouthValue: 0
   TypeOfGarnish: 0
+  IsJustWaterValue: 0
