+++ Assets/Prefabs/_Gameplay Prefabs/UI/Canvas - Game Over UI.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 93309652d3ecb754b8e6cd53807551dd
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
