+++ Assets/Prefabs/_Gameplay Prefabs/UI/MouseController UI_Canvas.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fad4cfc1c5bfbba43965f1b6d40fe289
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
