+++ Assets/Prefabs/_Gameplay Prefabs/UI/Score Display UI.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 19a98365947e2c948ab1a8327740215c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
