+++ Assets/Scripts/_Gameplay/BarktenderMovement.cs.meta
